backend/pyproject.toml
[project]
name = "ai-demo"
version = "1.0.0"
description = "AI Demo Project"
authors = [
    { name = "AI Demo Team" }
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0.post1",
    "pyyaml>=6.0.1"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "N", "W", "B"]
ignore = []

[tool.ruff.isort]
known-first-party = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v -s --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]
